Refactor actions: Consolidated createComponent and loadComponents imports into a single actions module, removed obsolete action files, and updated related components and hooks accordingly.
