--- packages/packages/core/editor-components/src/actions/load-components.ts
@@ -1,8 +1,0 @@
-import { __createAsyncThunk as createAsyncThunk } from '@elementor/store';
-
-import { apiClient } from '../api';
-
-export default createAsyncThunk( 'components/load', async () => {
-	const response = await apiClient.get();
-	return response;
-} );
