+++ packages/packages/core/editor-components/src/components/__tests__/components-tab.test.tsx
@@ -10,7 +10,7 @@ import {
 import { jest } from '@jest/globals';
 import { act, fireEvent, screen, waitFor } from '@testing-library/react';
 
-import loadComponents from '../../actions/load-components';
+import { loadComponents } from '../../actions';
 import { slice } from '../../store';
 import { ComponentSearch } from '../components-tab/component-search';
 import { ComponentItem } from '../components-tab/components-item';
@@ -288,24 +288,6 @@ describe( 'ComponentsTab', () => {
 			// Assert
 			expect( screen.getByText( 'Sorry, nothing matched' ) ).toBeInTheDocument();
 			expect( screen.getByText( 'Try something else.' ) ).toBeInTheDocument();
-		} );
-
-		it( 'should handle undefined components gracefully', () => {
-			// Arrange
-			act( () => {
-				dispatch( slice.actions.load( [] ) );
-			} );
-
-			// Act
-			renderWithStore(
-				<SearchProvider localStorageKey="test-search">
-					<ComponentsList />
-				</SearchProvider>,
-				store
-			);
-
-			// Assert
-			expect( screen.getByText( 'Text that explains that there are no Components yet.' ) ).toBeInTheDocument();
 		} );
 
 		it( 'should handle components with missing properties gracefully', () => {
